Rename rule to js-imports
