+++ src/index.ts
@@ -1,10 +1,10 @@
 'use strict'
 
-import noRelativePathImports from './rules/no-relative-path-imports'
+import noRelativePathImports from './rules/js-imports'
 import tsImports from './rules/ts-imports'
 
 // Import all rules in lib/rules
 module.exports.rules = {
   'ts-imports': tsImports,
-  'no-relative-path-imports': noRelativePathImports,
+  'js-imports': noRelativePathImports,
 }
